datamash: fix mode/antimode incorrect results for negative values
In version 1.4 and earlier, this resulted in incorrect values:

  $ echo -1 | datamash mode 1
  1.844674407371e+19

Reported by Renan Valieris <rvalieris@gmail.com> in
https://lists.gnu.org/archive/html/bug-datamash/2019-06/msg00001.html

* src/utils.c (mode_value): Fix type from size_t to long double.
* tests/datamash-tests-2.pl: Add test.
* NEWS: Mention fix.

+++ src/utils.c
@@ -381,7 +381,7 @@ mode_value ( const long double * const values, size_t n, enum MODETYPE type)
   long double last_value = values[0];
   size_t seq_size=1;
   size_t best_seq_size= (type==MODE)?1:SIZE_MAX;
-  size_t best_value = values[0];
+  long double best_value = values[0];
 
   for (size_t i=1; i<n; i++)
     {
